[Day12](P1,bugfix): left/right input are degrees
The input value for left and right turns are in degrees. As in
how many degrees the ship should turn.

But the value was errornously read as how many times the ship
should turn 90 degrees.

- R90 was read as "turn the ship right 90 times".
  If the ship was facing East, then it was later facing West.

- R90 should be read as "turn the ship 90 degrees to the right.
  If the ship was facing East, then it should be shanged to facing
  North.

The input value for left and right turns was errornously read as
the numbers of times to turn the ship.